fix(apps): Migrate and scope models to target app
Migrate county and constituency models from `common` to `county` app.

[Problem]
- Serializers from `common` app use undefined `county` app models.
- `common` app models host models belonging to `county` app.

[Note]
- Prefer to import `status` over specific constants for flexibility.
- Apply and unapply migrations after updates to app schemas.
- Delete the migrations dir for the `common` app.
