+++ recall_server/county/models.py
@@ -1,18 +1,41 @@
-# ruff: noqa: D101, DJ008
 """
 Custom Database models for the `county` Django app.
-
-Note:
-    Remove the noqa on line 1, once the following county models
-    have been implemented to re-enable ruff lint checks.
 """
 
 from django.db import models
+from recall_server.mps.models import MemberOfParliament
+from recall_server.polling_station.models import PollingStation
 
 
 class County(models.Model):
-    pass
+    """
+    Keep track of county and number of constituencies.
+    """
+
+    name = models.CharField(unique=True, max_length=40)
+    constituency_count = models.IntegerField(default=0)
+
+    def __str__(self):
+        return f"{self.name}: {self.constituency_count}"
 
 
 class Constituency(models.Model):
-    pass
+    """
+    Keep track of constituency.
+    """
+
+    name = models.CharField(unique=True, max_length=20)
+    registeredvoter_count = models.IntegerField(default=0)
+    mp = models.ForeignKey(
+        MemberOfParliament,
+        related_name="constituencies",
+        on_delete=models.CASCADE,
+    )
+    polling_station = models.ForeignKey(
+        PollingStation,
+        related_name="constituencies",
+        on_delete=models.CASCADE,
+    )
+
+    def __str__(self):
+        return f"{self.name} {self.mp} {self.polling_station}"
